E_Learn/requirements.txt
Django==5.1.7
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
Pillow==11.1.0
psycopg2-binary==2.9.10
Pylance